+++ jmbo/templatetags/jmbo_inclusion_tags.py
@@ -17,7 +17,7 @@ from django.utils.itercompat import is_iterable
 register = template.Library()
 
 
-# Base the inclusion_tag decorator on Django's default register tag. We want 
+# Base the inclusion_tag decorator on Django's default register tag. We want
 # to be able to dynamically compute a list of templates suitable for rendering.
 def inclusion_tag(register, context_class=Context, takes_context=False, name=None):
     def dec(func):
@@ -27,7 +27,7 @@ def inclusion_tag(register, context_class=Context, takes_context=False, name=Non
 
             def render(self, context):
                 resolved_args, resolved_kwargs = self.get_resolved_arguments(context)
-                
+
                 # Only this line has been changed from the default
                 # register_tag
                 file_name, _dict = func(*resolved_args, **resolved_kwargs)
@@ -66,13 +66,6 @@ def inclusion_tag(register, context_class=Context, takes_context=False, name=Non
         register.tag(function_name, compile_func)
         return func
     return dec
-
-
-@register.inclusion_tag('jmbo/inclusion_tags/content_list_gizmo.html', \
-        takes_context=True)
-def content_list_gizmo(context, object_list):
-    context.update({'object_list': object_list})
-    return context
 
 
 @inclusion_tag(register, takes_context=True)