+++ test_settings.py
@@ -14,20 +14,44 @@ DATABASES = {
     }
 }
 
-INSTALLED_APPS = [
+INSTALLED_APPS = (
+    'jmbo',
+    'photologue',
+    'category',
+    'likes',
+    'secretballot',
+    'pagination',
+    'publisher',
+    'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.comments',
     'django.contrib.contenttypes',
     'django.contrib.sites',
-#    'django.contrib.gis',
-#    'atlas',
-    'category',
-    'jmbo',
-    'photologue',
-    'secretballot',
-    'publisher',
     'south',
-]
+)
+
+ROOT_URLCONF = 'jmbo.tests.urls'
+
+MIDDLEWARE_CLASSES = (
+    'django.middleware.common.CommonMiddleware',
+    'django.contrib.sessions.middleware.SessionMiddleware',
+    'django.middleware.csrf.CsrfViewMiddleware',
+    'django.contrib.auth.middleware.AuthenticationMiddleware',
+    'django.contrib.messages.middleware.MessageMiddleware',
+    'likes.middleware.SecretBallotUserIpUseragentMiddleware',
+    'pagination.middleware.PaginationMiddleware',
+)
+
+TEMPLATE_CONTEXT_PROCESSORS = (
+    'django.contrib.auth.context_processors.auth',
+    'django.core.context_processors.debug',
+    'django.core.context_processors.i18n',
+    'django.core.context_processors.media',
+    'django.core.context_processors.static',
+    'django.core.context_processors.tz',
+    'django.core.context_processors.request',
+    'django.contrib.messages.context_processors.messages',
+)
 
 SITE_ID = 1