+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ next
 #. Add `Clone this item` button to change forms.
 #. Deprecate gizmo, "wide" template, Pin class.
 #. Deprecate own class based generic views in favour of Django's equivalent.
+#. Deprecate views related to show objects per category. `jmbio-foundry` offers a much more powerful solution and scales better.
 
 1.1.5
 -----