+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+1.1.6
+-----
+#. API now includes image detail url.
+#. URL pattern to resolve detail page through category.
+
 1.1.5
 -----
 #. Ignore result of celery tasks as appropriate.